requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-groq
transformers
python-dotenv
ipykernel
youtube-transcript-api
langchain-chroma
chromadb
sentence-transformers
duckduckgo-search 
langchain_experimental
ddgs
langgraph
streamlit
langgraph-checkpoint-sqlite